feat: Update service request and model to include new fields, enhance chambre retrieval, and adjust routes for employee-specific access

+++ routes/api.php
@@ -132,21 +132,15 @@ Route::prefix('v1')->group(function () {
             Route::put('/employes/{slug}', [EmployeController::class, 'update']);
             Route::delete('/employes/{slug}', [EmployeController::class, 'destroy']);
 
-
             Route::post('/entreprises', [EntrepriseController::class, 'store']);
             Route::get('/owner/entreprises', [EntrepriseController::class, 'findByUser']);
-            Route::get('/entreprises/{slug}', [EntrepriseController::class, 'show']);
             Route::post('/entreprises/{slug}', [EntrepriseController::class, 'update']);
-
             Route::delete('/entreprises/{slug}', [EntrepriseController::class, 'destroy']);
 
             Route::get('/owner/riads', [RiadController::class, 'findByUser']);
-            Route::post('/riads', [RiadController::class, 'store']);
-            Route::get('/riads/{slug}', [RiadController::class, 'show']);
-            Route::put('/riads/{slug}', [RiadController::class, 'update']);
-            Route::delete('/riads/{slug}', [RiadController::class, 'destroy']);
-
-
+            Route::post('/owner/riads', [RiadController::class, 'store']);
+            Route::put('/owner/riads/{slug}', [RiadController::class, 'update']);
+            Route::delete('/owner/riads/{slug}', [RiadController::class, 'destroy']);
         });
